+++ src/panos_topo.py
@@ -7,20 +7,35 @@ import geopandas as gpd
 from collections import deque
 from utils.geo_helper import gdf_to_geojson
 
-from utils.pickle_helper import PickleSaver
 from utils.unionFind import UnionFind
-from utils.azimuth_helper import azimuth_diff
+from utils.geo_helper import gdf_to_geojson
+from utils.pickle_helper import PickleSaver
 from utils.geo_plot_helper import map_visualize
 from utils.log_helper import LogHelper, logbook
 from pano_base import pano_dict_to_gdf, extract_gdf_roads_from_key_pano, extract_gdf_panos_from_key_pano
 
 
-logger = LogHelper(log_dir="../log", log_name='pano_topo.log').make_logger(level=logbook.INFO)
-
 #%%
 
 def query_edge_by_node(id, df_topo):
     return df_topo.query( "src == @id or dst == @id" )
+
+
+def azimuth_diff(a, b, unit='radian'):
+    """calcaluate the angle diff between two azimuth
+    Args:
+        a ([type]): Unit: degree
+        b ([type]): Unit: degree
+        unit(string): radian or degree
+    Returns:
+        [type]: [description]
+    """
+    diff = abs(a-b)
+
+    if diff > 180:
+        diff = 360-diff
+
+    return diff if unit =='degree' else diff*math.pi/180
 
 
 def plot_node_connections(node, *args, **kwargs):
@@ -141,21 +156,22 @@ def build_graph(df_topo):
     return graph
 
 
-def bfs(node, df_topo, direction=True, visited=set(), plot=True, similar_threds=.9):
+def bfs(node, df_topo, direction=True, visited=set(), plot=True, similar_threds=.9, logger=None):
     if node not in df_topo.index:
         return []
     
     queue, res = deque([node]), []
-    
     while queue:
         cur_pid = queue.popleft()
         if cur_pid not in df_topo.index:
             continue
-        logger.info(f"{cur_pid}, {'forword' if direction else 'backward'}: {df_topo.loc[cur_pid].index.values.tolist()}")
 
+        if logger is not None:
+            logger.info(f"{cur_pid}, {'forword' if direction else 'backward'}: {df_topo.loc[cur_pid].index.values.tolist()}")
         for nxt_pid, nxt in df_topo.loc[cur_pid].iterrows():
             connector = (cur_pid, nxt_pid) if direction else (nxt_pid, cur_pid)
-            logger.debug(f"{connector}, {nxt_pid}, {nxt_pid not in df_topo.index}")
+            if logger is not None:
+                logger.debug(f"{connector}, {nxt_pid}, {nxt_pid not in df_topo.index}")
             
             if nxt['similarity'] < similar_threds or connector in visited:
                 continue
@@ -166,9 +182,9 @@ def bfs(node, df_topo, direction=True, visited=set(), plot=True, similar_threds=
                 res.append(nxt['rid'])
             
             queue.append(nxt_pid)
-            break
+            visited.add(connector)
 
-        visited.add(connector)
+            break
     
     if plot and len(res) > 0:
         map_visualize( gdf_roads.loc[res] )
@@ -176,7 +192,7 @@ def bfs(node, df_topo, direction=True, visited=set(), plot=True, similar_threds=
     return res
 
 
-def bidirection_bfs(pid, df_topo, df_topo_prev, visited=set(), plot=True):
+def bidirection_bfs(pid, df_topo, df_topo_prev, visited=set(), plot=True, logger=None):
     """[summary]
 
     Args:
@@ -193,11 +209,12 @@ def bidirection_bfs(pid, df_topo, df_topo_prev, visited=set(), plot=True):
     # assert pid in df_topo.index, f"check the pid input: {pid}"
     
     if pid not in df_topo.index and pid not in df_topo_prev.index:
-        logger.warning(f"{pid} not in df_topo")
+        if logger is not None:
+            logger.warning(f"{pid} not in df_topo")
         return []
 
-    rids_0 = bfs(pid, df_topo_prev, False, visited, False)
-    rids_1 = bfs(pid, df_topo, True, visited, False)
+    rids_0 = bfs(pid, df_topo_prev, False, visited, False, logger=logger)
+    rids_1 = bfs(pid, df_topo, True, visited, False, logger=logger)
     rids = rids_0[::-1] + rids_1
    
     if rids and plot:
@@ -209,7 +226,7 @@ def bidirection_bfs(pid, df_topo, df_topo_prev, visited=set(), plot=True):
     return rids
 
 
-def combine_rids(gdf_base, gdf_roads, plot=True):
+def combine_rids(gdf_base, gdf_roads, plot=True, logger=None):
     """Combining rids based on the amuzith similarity.
 
     Args:
@@ -221,7 +238,6 @@ def combine_rids(gdf_base, gdf_roads, plot=True):
     df_topo, df_topo_prev = get_topo_from_gdf_pano(gdf_base)
 
     res            = {}
-    uf             = UnionFind(gdf_roads.index.tolist())
     rid_2_start    = {}
     # 起点和终点都是一样的路段不适宜作为遍历的起始点
     queue          = deque(gdf_roads.query('src != dst').index.tolist())
@@ -233,20 +249,18 @@ def combine_rids(gdf_base, gdf_roads, plot=True):
         cur_pid = cur['src']
 
         if (cur_pid, cur['dst']) in visited:
-            logger.debug(f"skip {cur.name}")
+            if logger is not None:
+                logger.debug(f"skip {cur.name}")
             continue
         
-        rids = bidirection_bfs(cur_pid, df_topo, df_topo_prev, visited, plot=False)
+        rids = bidirection_bfs(cur_pid, df_topo, df_topo_prev, visited, plot=False, logger=logger)
         if not rids:
             continue
 
-        # origin
-        for i in range(len(rids)-1):
-            uf.connect(rids[i], rids[i+1])
-        # new
         for i in rids:
             if i in rid_2_start:
-                logger.warning(f"{i} exist in rid_2_start")
+                if logger is not None:
+                    logger.warning(f"{i} exist in rid_2_start")
             rid_2_start[i] = rid_2_start.get(i, set())
             rid_2_start[i].add(rids[0])
         
@@ -256,14 +270,15 @@ def combine_rids(gdf_base, gdf_roads, plot=True):
             res[rids[0]].append(rids)
             dulplicate_src.append(rids[0])
             print(f"dulplicate: {rids[0]}")
-        
-        logger.info(f"visited {cur_pid}/{cur.name}, links: {rids}")
+
+        if logger is not None:
+            logger.debug(f"visited {cur_pid}/{cur.name}, links: {rids}")
 
     if plot:
         fig, ax = map_visualize(gdf_roads, scale=.05)
         gdf_roads.merge( pd.DataFrame(visited, columns=['src', 'dst']), on=['src', 'dst']).plot(ax=ax)
 
-    return res, uf, rid_2_start
+    return res, rid_2_start
 
 
 def debug_shrink_links():
@@ -309,7 +324,7 @@ def get_panos_by_rids(rids, gdf_roads, gdf_panos=None, plot=False):
     return gdf.reset_index()
 
 
-def get_trajectory_by_rid(rid, uf, traj_rid_lst, gdf_roads, plot=True):
+def get_trajectory_by_rid(rid, rid_2_start, traj_rid_lst, gdf_roads, plot=True):
     """选择 rid 所在的轨迹，也可以进行可视化
 
     Args:
@@ -319,10 +334,14 @@ def get_trajectory_by_rid(rid, uf, traj_rid_lst, gdf_roads, plot=True):
     Returns:
         [type]: [description]
     """
-    if rid not in uf.father:
+    if rid not in rid_2_start:
         return []
     
-    rids = traj_rid_lst[uf.find(rid)]
+    rids = []
+    for src in rid_2_start[rid]:
+        values = traj_rid_lst[src]
+        rids += values
+
     for values in rids:
         if rid in values:
             if plot:
@@ -330,7 +349,8 @@ def get_trajectory_by_rid(rid, uf, traj_rid_lst, gdf_roads, plot=True):
                 roads.loc[:, 'order'] = range(roads.shape[0])
                 fig, ax = map_visualize(roads, color='black')
                 roads.plot(column='order', legend=True, ax=ax, linewidth=3)
-                
+            
+            # FIXME 存在一个节点两个分支都有rid的情况，但目前展示不处理
             return values
     
     return []
@@ -338,6 +358,8 @@ def get_trajectory_by_rid(rid, uf, traj_rid_lst, gdf_roads, plot=True):
 
 #%%
 if __name__ == '__main__':
+    logger = LogHelper(log_dir="../log", log_name='pano_topo.log').make_logger(level=logbook.INFO)
+
     pickler   = PickleSaver()
     # pano_dict = pickler.read('../cache/pano_dict_lxd.pkl')
     pano_dict = pickler.read('../cache/pano_dict_futian.pkl')
@@ -351,6 +373,10 @@ if __name__ == '__main__':
     query_edge_by_node('09005700121709091548023739Y', df_topo)
 
 
+    """ plot node and its conncetions """
+    plot_node_connections('09005700121709131151514569Y')
+
+
     """ check for single direaction bfs """
     pid = '09005700121709031455036592S'
     rids = bfs(pid, df_topo, True)
@@ -362,41 +388,29 @@ if __name__ == '__main__':
 
 
     """ check for bidirection bfs """
-    # 创科路北行
-    # bidirection_bfs('09005700121709091038208679Y', df_topo, df_topo_prev)
-    # 创科路南行
-    # bidirection_bfs('09005700121709091041425059Y', df_topo, df_topo_prev)
-    # 打石一路东行
-    # bidirection_bfs('09005700121709091548023739Y', df_topo, df_topo_prev)
-    # 打石一路西行
-    bidirection_bfs('09005700121709091542295739Y', df_topo, df_topo_prev)
-    # 特殊案例
-    bidirection_bfs('09005700121709091542314649Y', df_topo, df_topo_prev, set(), True)
-
-    
-    """ plot node and its conncetions """
-    plot_node_connections('09005700121709131151514569Y')
+    # bidirection_bfs('09005700121709091038208679Y', df_topo, df_topo_prev) # 创科路北行
+    # bidirection_bfs('09005700121709091041425059Y', df_topo, df_topo_prev) # 创科路南行
+    # bidirection_bfs('09005700121709091548023739Y', df_topo, df_topo_prev) # 打石一路东行
+    bidirection_bfs('09005700121709091542295739Y', df_topo, df_topo_prev, logger=logger)# 打石一路西行
+    bidirection_bfs('09005700121709091542314649Y', df_topo, df_topo_prev, set(), True, logger=logger) # 特殊案例
 
 
     """" combine rids """
-    traj_rid_lst, rid_uf, rid_2_start = combine_rids(gdf_base, gdf_roads, plot=True)
+    traj_rid_lst, rid_2_start = combine_rids(gdf_base, gdf_roads, plot=False, logger=logger)
 
 
     """ 根据rid 查询轨迹 """
     # 留仙洞
-    get_trajectory_by_rid("988acb-1732-52e3-a58a-36eec3", rid_uf, gdf_roads)
-    # fitian 
-    df = get_trajectory_by_rid('550a27-40c5-f0d3-5717-a1907d', rid_uf, gdf_roads)
-    from utils.geo_helper import gdf_to_geojson
-    gdf_to_geojson(df, '../cache/panos_for_test')
+    get_trajectory_by_rid("988acb-1732-52e3-a58a-36eec3", rid_2_start, traj_rid_lst, gdf_roads)
+    # futian 
+    rid = '550a27-40c5-f0d3-5717-a1907d' # 金田路福田地铁站附近
+    get_trajectory_by_rid(rid, rid_2_start, traj_rid_lst, gdf_roads) 
 
-#%%
+    rid = 'edbf2d-e2f3-703f-4b9f-9d6819' # 深南大道-市民中心-东行掉头
+    get_trajectory_by_rid(rid, rid_2_start, traj_rid_lst, gdf_roads)
 
-traj_rid_lst, rid_uf, rid_2_start = combine_rids(gdf_base, gdf_roads, plot=True)
+    rid = 'd51f52-4ab6-cba6-dc4f-2fdf73' # 深南大道/益田路立交-东侧
+    get_trajectory_by_rid(rid, rid_2_start, traj_rid_lst, gdf_roads)
 
-# get_trajectory_by_rid('586fa0-4623-8530-bcea-12e41d', rid_uf, traj_rid_lst, gdf_roads, plot=True)
-
-get_trajectory_by_rid('52fec5-aae7-3da8-56ee-dfca06', rid_uf, traj_rid_lst, gdf_roads, plot=True)
-
-
-# %%
+    rid = '514cba-89ea-b8d6-3de2-15f9ac' # 深南大道/益田路立交-西侧
+    get_trajectory_by_rid(rid, rid_2_start, traj_rid_lst, gdf_roads)